refactor: 뷰 리팩터링
