+++ src/main/java/janggi/domain/game/SetupOption.java
@@ -1,0 +1,6 @@
+package janggi.domain.game;
+
+public enum SetupOption {
+
+    INNER_SETUP, OUTER_SETUP, RIGHT_SETUP, LEFT_SETUP
+}
